Update FWIpart2.md
Nice work. Made edits and added comments in critic markup. My main critique is that we should bring in the numerics earlier and one thought is to explain the adjoint by forming explicit matrices for a small 1D problem. I also strongly suggest to add gradient descent here so we have our hands free in part 3. Finally, I would drop (as in part 1) the operator notation and leave that for part 3. Needs some work but we are getting there.